--- a/package.json
+++ b/package.json
@@ -39,6 +39,7 @@
     "@casl/mongoose": "^5.0.0",
     "@ucast/mongo2js": "^1.3.3",
     "amqplib": "^0.8.0",
+    "apollo-datasource-mongodb": "^0.5.4",
     "apollo-datasource-rest": "^3.5.2",
     "apollo-server-caching": "^3.3.0",
     "apollo-server-core": "^2.25.2",