--- a/src/models/form.model.ts
+++ b/src/models/form.model.ts
@@ -185,7 +185,7 @@ export const Form: FormModel = mongoose.model<Form, FormModel>('Form', schema);
 export class FormDataSource extends MongoDataSource<FormDocument> {
   /**
    * Get all forms
-   * 
+   *
    * @returns Form data.
    */
   async getForms() {
@@ -194,7 +194,7 @@ export class FormDataSource extends MongoDataSource<FormDocument> {
 
   /**
    * Get form detail by id
-   * 
+   *
    * @param id is user id
    * @returns Single Form data.
    */