# Description
- have a main resource, linked to other resources
- have many different records for these resources
- make sure that the records are connected between them
- make sure that some fields are using Reference Data
- create a layout in a grid, that would consume records from this main resource
- it should take a really long time to query the records

## Type of change
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Improvement (refactor or addition to existing functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

# How Has This Been Tested?
please create form with resource/resources field and add dashboard page for with grid widget

# Result
- before integrated datasource I was get 400 records in the 50 to 55 seconds and 200 records in the 25 to 30 seconds
- after integarting datasource I am getting 400 recods in the 10 to 12 seconds and 200 records in the 5 to 7 seconds
- Also I have checked with dashboard in the frontend, before integrated 100 records take 45 to 50 second to display and now 5 to 8 seconds to display record in the dashboard

# Checklist:
- [x] * My code follows the style guidelines of this project
- [x] * Linting does not generate new warnings
- [x] * I have performed a self-review of my own code
- [x] * I have commented my code, particularly in hard-to-understand areas
- [x] * I have put JSDoc comment in all required places
- [x] * My changes generate no new warnings
- [x] * I have included screenshots describing my changes if relevant
- [x] * I have selected labels in the Pull Request, according to the changes with code brings
- [ ] I have made corresponding changes to the documentation ( if required )
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules